refactor : #16 파일 정리
