--- src/components/mypage/main-theme.css
@@ -1,126 +1,0 @@
-body, html {
-    margin: 0;
-    padding: 0;
-    font-family: 'Noto Sans KR', sans-serif;
-    background-color: #000;
-    color: #0f0;
-    height: 100%;
-    box-sizing: border-box;
-}
-*, *:before, *:after {
-    box-sizing: inherit;
-}
-.container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 2rem;
-}
-header {
-    background-color: rgba(0, 255, 0, 0.1);
-    padding: 1rem;
-    position: fixed;
-    width: 100%;
-    top: 0;
-    z-index: 1000;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-.logo {
-    font-size: 1.5rem;
-    font-weight: bold;
-    color: #0f0;
-    text-shadow: 0 0 10px #0f0;
-    text-decoration: none;
-}
-.buttons {
-    display: flex;
-    gap: 1rem;
-}
-.buttons button {
-    color: #0f0;
-    background-color: #000;
-    border: 1px solid #0f0;
-    padding: 0.5rem 1rem;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-}
-.buttons button:hover {
-    background-color: #0f0;
-    color: #000;
-}
-h1, h2 {
-    text-align: center;
-    margin-top: 5rem;
-    text-shadow: 0 0 15px #0f0;
-}
-.popular-themes {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: 2rem;
-    margin-top: 2rem;
-}
-.theme-card {
-    background-color: rgba(0, 255, 0, 0.05);
-    border-radius: 10px;
-    padding: 1rem;
-    text-align: center;
-    transition: all 0.3s ease;
-    display: flex;
-    flex-direction: column;
-}
-.theme-card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 5px 15px rgba(0, 255, 0, 0.3);
-}
-.theme-image {
-    width: 100%;
-    height: 200px;
-    object-fit: cover;
-    border-radius: 5px;
-    margin-bottom: 1rem;
-}
-.theme-details {
-    flex-grow: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-.theme-title {
-    font-size: 1.2rem;
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-}
-.theme-location, .theme-genre, .theme-difficulty {
-    font-size: 0.9rem;
-    margin-bottom: 0.3rem;
-}
-.theme-rating {
-    display: flex;
-    align-items: center;
-    margin-bottom: 0.5rem;
-}
-.star {
-    color: #ff0;
-    font-size: 1.2rem;
-    margin-right: 0.2rem;
-}
-.book-button {
-    background-color: #0f0;
-    color: #000;
-    border: none;
-    padding: 0.5rem 1rem;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    border-radius: 5px;
-    font-size: 1rem;
-    text-align: center;
-    text-decoration: none;
-    display: block;
-}
-.book-button:hover {
-    background-color: #000;
-    color: #0f0;
-    border: 1px solid #0f0;
-}
