+++ src/router.js
@@ -1,13 +1,13 @@
 // import Vue from "vue";
 import {createRouter, createWebHistory} from "vue-router";
-import SignUp from './components/signup/SignUp.vue';
+import SignUp from './components/auth/SignUp.vue';
 import MyPage from "@/views/MyPage.vue";
-import Login from "@/components/login/Login.vue";
+import Login from "@/components/auth/Login.vue";
 import UserInfo from "@/components/mypage/UserInfo.vue"
 import ReservationInfo from "@/components/mypage/ReservationInfo.vue";
 import FollowInfo from "@/components/mypage/FollowInfo.vue";
-import MainTheme from '@/components/MainTheme.vue'
-import SearchStore from '@/components/SearchStore.vue'
+import MainTheme from '@/views/MainTheme.vue'
+import SearchStore from '@/views/SearchStore.vue'
 import Reservation from "@/views/Reservation.vue";
 import bookingInfo from '@/components/payment/BookingInfo.vue'
 import bookingForm from '@/components/payment/BookingForm.vue'
@@ -16,7 +16,7 @@ import paySuccess from './components/payment/Home.vue';
 import payCancel from './components/payment/Cancel.vue';
 import payFail from './components/payment/Fail.vue';
 import ReviewList from '@/components/review/ReviewList.vue'
-import ThemeInfo from '@/components/ThemeInfo.vue'
+import ThemeInfo from '@/views/ThemeInfo.vue'
 
 export const router = createRouter({
     history: createWebHistory(),
